Move current Gmp class to gmp.protocol.v7
With this change we can support different gmp protocol versions in
gvm-tools in future.
