+++ gmp/__init__.py
@@ -19,7 +19,7 @@
 Main module for gvm-tools
 """
 from gmp.error import GmpError
-from gmp.gmp import Gmp
+from gmp.protocol.v7 import Gmp
 
 VERSION = (2, 0, 0, 'dev', 1)
